refactor: Change file name ContractInfo to UsageHistory
