+++ Views/User/Admin/ad_UsageHistory.ejs
@@ -15,7 +15,7 @@
     </header>
     <section>
       <div class="container">
-        <h2 class="mt-5 mb-3">Contract Info list</h2>
+        <h2 class="mt-5 mb-3">Usage History</h2>
         <div class="table-responsive">
           <table class="table table-striped ">
             <thead class="thead-dark">
@@ -53,7 +53,7 @@
                 <td class="col-2" style="padding-top: 20px;">Not Currently Used</td>
                 <% } %>
                 <td class="col-1" style="padding:6px 0;">
-                  <form action="/Admin/ContractInfo/INFO" method="POST">
+                  <form action="/Admin/UsageHistory/INFO" method="POST">
                     <input id="reqID" name="reqID" type="hidden" value="<%= items['item' + i].reqID %>">
                     <button class="btn btn-primary" style="margin: 0; padding: 8px 12px;">INFO</button>
                   </form>
