+++ Views/nav_Bar.ejs
@@ -16,7 +16,7 @@
                     <a class="dropdown-item" href="/Admin/RequestEnroll" id="account_ctrl">Warehouse Enroll Request</a>
                     <a class="dropdown-item" href="/Admin/RequestIoT">IoT Service Request</a>
                     <a class="dropdown-item" href="/Admin/RequestBuy">Warehouse Purchase Request</a>
-                    <a class="dropdown-item" href="/Admin/ContractInfo">Contract Information</a>
+                    <a class="dropdown-item" href="/Admin/UsageHistory">Usage History</a>
                   </div>
                 </div>
               </div>
@@ -25,7 +25,7 @@
               <li class="nav-item"><a class="nav-link js-scroll-trigger" href="/Admin/RequestEnroll">Warehouse Enroll Request</a></li>
               <li class="nav-item"><a class="nav-link js-scroll-trigger" href="/Admin/RequestIoT">IoT Service Request</a></li>
               <li class="nav-item"><a class="nav-link js-scroll-trigger" href="/Admin/RequestBuy">Warehouse Purchase Request</a></li>
-              <li class="nav-item"><a class="nav-link js-scroll-trigger" href="/Admin/ContractInfo">Contract Information</a></li>
+              <li class="nav-item"><a class="nav-link js-scroll-trigger" href="/Admin/UsageHistory">Usage History</a></li>
             </div>
             <li class="nav-item"><a href="/User/Show" class="nav-link js-scroll-trigger">My Info</a></li>
           <% } else if(session.type == 'buyer') { %>
